+++ structs.go
@@ -9,7 +9,7 @@ type Device struct {
 		Time   string `json:"time"`
 		Sid    int64  `json:"sid"`
 	} `json:"ips,omitempty"`
-	Did             int     `json:"did,omitempty"`
+	Did             int64   `json:"did,omitempty"`
 	Sid             int64   `json:"sid,omitempty"`
 	Hostname        string  `json:"hostname,omitempty"`
 	Time            int64   `json:"time,omitempty"`
@@ -22,46 +22,46 @@ type Device struct {
 
 type DeviceInfo struct {
 	DeviceInfo []struct {
-		Did             int `json:"did"`
-		SimilarityScore int `json:"similarityScore"`
+		Did             int64 `json:"did"`
+		SimilarityScore int64 `json:"similarityScore"`
 		GraphData       []struct {
 			Time  int64 `json:"time"`
-			Count int   `json:"count"`
+			Count int64 `json:"count"`
 		} `json:"graphData"`
 		Info struct {
-			TotalUsed            int `json:"totalUsed"`
-			TotalServed          int `json:"totalServed"`
-			TotalDevicesAndPorts int `json:"totalDevicesAndPorts"`
+			TotalUsed            int64 `json:"totalUsed"`
+			TotalServed          int64 `json:"totalServed"`
+			TotalDevicesAndPorts int64 `json:"totalDevicesAndPorts"`
 			DevicesAndPorts      []struct {
 				DeviceAndPort string `json:"deviceAndPort"`
-				Size          int    `json:"size"`
+				Size          int64  `json:"size"`
 			} `json:"devicesAndPorts"`
 			PortsUsed []struct {
 				Port      interface{} `json:"port"`
-				Size      int         `json:"size"`
+				Size      int64       `json:"size"`
 				FirstTime int64       `json:"firstTime"`
 			} `json:"portsUsed"`
 			PortsServed []struct {
 				Port interface{} `json:"port"`
-				Size int         `json:"size"`
+				Size int64       `json:"size"`
 			} `json:"portsServed"`
 			DevicesUsed []struct {
 				Did       interface{} `json:"did"`
-				Size      int         `json:"size"`
+				Size      int64       `json:"size"`
 				FirstTime int64       `json:"firstTime,omitempty"`
 			} `json:"devicesUsed"`
 			DevicesServed []struct {
 				Did  interface{} `json:"did"`
-				Size int         `json:"size"`
+				Size int64       `json:"size"`
 			} `json:"devicesServed"`
 		} `json:"info"`
 	} `json:"deviceInfo"`
 }
 
 type SimilarDevice struct {
-	Did        int    `json:"did"`
+	Did        int64  `json:"did"`
 	Objecttype string `json:"objecttype"`
-	Score      int    `json:"score"`
+	Score      int64  `json:"score"`
 	IP         string `json:"ip"`
 	Ips        []struct {
 		IP     string `json:"ip"`
@@ -76,29 +76,29 @@ type SimilarDevice struct {
 	Os        string `json:"os,omitempty"`
 	Typename  string `json:"typename"`
 	Typelabel string `json:"typelabel"`
-	Priority  int    `json:"priority,omitempty"`
+	Priority  int64  `json:"priority,omitempty"`
 }
 
 type DeviceMetric struct {
 	Time   string `json:"time"`
 	TimeMS int64  `json:"timems"`
-	Size   int    `json:"size"`
-	In     int    `json:"in"`
-	Out    int    `json:"out"`
+	Size   int64  `json:"size"`
+	In     int64  `json:"in"`
+	Out    int64  `json:"out"`
 }
 
 type Model struct {
 	Name  string `json:"name"`
-	Pid   int    `json:"pid"`
-	Phid  int    `json:"phid"`
+	Pid   int64  `json:"pid"`
+	Phid  int64  `json:"phid"`
 	UUID  string `json:"uuid"`
 	Logic struct {
 		Data    []interface{} `json:"data"`
 		Type    string        `json:"type"`
-		Version int           `json:"version"`
+		Version int64         `json:"version"`
 	} `json:"logic"`
-	Throttle        int  `json:"throttle"`
-	SharedEndpoints bool `json:"sharedEndpoints"`
+	Throttle        int64 `json:"throttle"`
+	SharedEndpoints bool  `json:"sharedEndpoints"`
 	Actions         struct {
 		Alert    bool `json:"alert"`
 		Antigena struct {
@@ -110,7 +110,7 @@ type Model struct {
 		SetType     interface{} `json:"setType"`
 	} `json:"actions"`
 	Tags        []string `json:"tags"`
-	Interval    int      `json:"interval"`
+	Interval    int64    `json:"interval"`
 	Sequenced   bool     `json:"sequenced"`
 	Active      bool     `json:"active"`
 	Modified    string   `json:"modified"`
@@ -120,9 +120,9 @@ type Model struct {
 		Tags struct {
 		} `json:"tags"`
 		Type    string `json:"type"`
-		Version int    `json:"version"`
+		Version int64  `json:"version"`
 	} `json:"activeTimes"`
-	Priority      int           `json:"priority"`
+	Priority      int64         `json:"priority"`
 	AutoUpdatable bool          `json:"autoUpdatable"`
 	AutoUpdate    bool          `json:"autoUpdate"`
 	AutoSuppress  bool          `json:"autoSuppress"`
@@ -140,18 +140,18 @@ type Model struct {
 		Active   bool   `json:"active"`
 		Message  string `json:"message"`
 		By       string `json:"by"`
-		Phid     int    `json:"phid"`
+		Phid     int64  `json:"phid"`
 	} `json:"history"`
 	Message string `json:"message"`
-	Version int    `json:"version"`
+	Version int64  `json:"version"`
 }
 
 type ModelComponent struct {
-	Cid       int `json:"cid"`
-	Chid      int `json:"chid"`
-	Mlid      int `json:"mlid"`
-	Threshold int `json:"threshold"`
-	Interval  int `json:"interval"`
+	Cid       int64 `json:"cid"`
+	Chid      int64 `json:"chid"`
+	Mlid      int64 `json:"mlid"`
+	Threshold int64 `json:"threshold"`
+	Interval  int64 `json:"interval"`
 	Logic     struct {
 		Data    interface{} `json:"data"`
 		Version string      `json:"version"`
@@ -162,22 +162,22 @@ type ModelComponent struct {
 
 type ModelBreach struct {
 	CreationTime int64 `json:"creationTime"`
-	CommentCount int   `json:"commentCount"`
+	CommentCount int64 `json:"commentCount"`
 	Pbid         int64 `json:"pbid"`
 	Time         int64 `json:"time"`
 	Model        struct {
 		Then struct {
 			Name  string `json:"name"`
-			Pid   int    `json:"pid"`
-			Phid  int    `json:"phid"`
+			Pid   int64  `json:"pid"`
+			Phid  int64  `json:"phid"`
 			UUID  string `json:"uuid"`
 			Logic struct {
 				Data    interface{} `json:"data"`
 				Type    string      `json:"type"`
-				Version int         `json:"version"`
+				Version int64       `json:"version"`
 			} `json:"logic"`
-			Throttle        int  `json:"throttle"`
-			SharedEndpoints bool `json:"sharedEndpoints"`
+			Throttle        int64 `json:"throttle"`
+			SharedEndpoints bool  `json:"sharedEndpoints"`
 			Actions         struct {
 				Alert    bool `json:"alert"`
 				Antigena struct {
@@ -189,7 +189,7 @@ type ModelBreach struct {
 				SetType     bool `json:"setType"`
 			} `json:"actions"`
 			Tags        []string `json:"tags"`
-			Interval    int      `json:"interval"`
+			Interval    int64    `json:"interval"`
 			Sequenced   bool     `json:"sequenced"`
 			Active      bool     `json:"active"`
 			ReadOnly    bool     `json:"readOnly"`
@@ -200,9 +200,9 @@ type ModelBreach struct {
 				Tags struct {
 				} `json:"tags"`
 				Type    string `json:"type"`
-				Version int    `json:"version"`
+				Version int64  `json:"version"`
 			} `json:"activeTimes"`
-			Priority      int           `json:"priority"`
+			Priority      int64         `json:"priority"`
 			AutoUpdatable bool          `json:"autoUpdatable"`
 			AutoUpdate    bool          `json:"autoUpdate"`
 			AutoSuppress  bool          `json:"autoSuppress"`
@@ -213,20 +213,20 @@ type ModelBreach struct {
 			} `json:"created"`
 			Edited struct {
 			} `json:"edited"`
-			Version int `json:"version"`
+			Version int64 `json:"version"`
 		} `json:"then"`
 		Now struct {
 			Name  string `json:"name"`
-			Pid   int    `json:"pid"`
-			Phid  int    `json:"phid"`
+			Pid   int64  `json:"pid"`
+			Phid  int64  `json:"phid"`
 			UUID  string `json:"uuid"`
 			Logic struct {
 				Data    interface{} `json:"data"`
 				Type    string      `json:"type"`
-				Version int         `json:"version"`
+				Version int64       `json:"version"`
 			} `json:"logic"`
-			Throttle        int  `json:"throttle"`
-			SharedEndpoints bool `json:"sharedEndpoints"`
+			Throttle        int64 `json:"throttle"`
+			SharedEndpoints bool  `json:"sharedEndpoints"`
 			Actions         struct {
 				Alert    bool `json:"alert"`
 				Antigena struct {
@@ -238,7 +238,7 @@ type ModelBreach struct {
 				SetType     bool `json:"setType"`
 			} `json:"actions"`
 			Tags        []string `json:"tags"`
-			Interval    int      `json:"interval"`
+			Interval    int64    `json:"interval"`
 			Sequenced   bool     `json:"sequenced"`
 			Active      bool     `json:"active"`
 			ReadOnly    bool     `json:"readOnly"`
@@ -249,9 +249,9 @@ type ModelBreach struct {
 				Tags struct {
 				} `json:"tags"`
 				Type    string `json:"type"`
-				Version int    `json:"version"`
+				Version int64  `json:"version"`
 			} `json:"activeTimes"`
-			Priority      int           `json:"priority"`
+			Priority      int64         `json:"priority"`
 			AutoUpdatable bool          `json:"autoUpdatable"`
 			AutoUpdate    bool          `json:"autoUpdate"`
 			AutoSuppress  bool          `json:"autoSuppress"`
@@ -262,24 +262,24 @@ type ModelBreach struct {
 			} `json:"created"`
 			Edited struct {
 			} `json:"edited"`
-			Version int `json:"version"`
+			Version int64 `json:"version"`
 		} `json:"now"`
 	} `json:"model"`
 	TriggeredComponents []struct {
 		Time      int64 `json:"time"`
 		Cbid      int64 `json:"cbid"`
-		Cid       int   `json:"cid"`
-		Chid      int   `json:"chid"`
-		Size      int   `json:"size"`
-		Threshold int   `json:"threshold"`
-		Interval  int   `json:"interval"`
+		Cid       int64 `json:"cid"`
+		Chid      int64 `json:"chid"`
+		Size      int64 `json:"size"`
+		Threshold int64 `json:"threshold"`
+		Interval  int64 `json:"interval"`
 		Metric    struct {
-			Mlid  int    `json:"mlid"`
+			Mlid  int64  `json:"mlid"`
 			Name  string `json:"name"`
 			Label string `json:"label"`
 		} `json:"metric"`
 		TriggeredFilters []struct {
-			Cfid       int    `json:"cfid"`
+			Cfid       int64  `json:"cfid"`
 			ID         string `json:"id"`
 			FilterType string `json:"filterType"`
 			Arguments  struct {
@@ -293,7 +293,7 @@ type ModelBreach struct {
 	} `json:"triggeredComponents"`
 	Score  float64 `json:"score"`
 	Device struct {
-		Did             int     `json:"did"`
+		Did             int64   `json:"did"`
 		Objecttype      string  `json:"objecttype"`
 		Hostname        string  `json:"hostname"`
 		IP              string  `json:"ip"`
@@ -316,11 +316,11 @@ type Enum struct {
 
 type EndpointDetails struct {
 	Hostname   string `json:"hostname"`
-	Dgascore   int    `json:"dgascore"`
+	Dgascore   int64  `json:"dgascore"`
 	Firsttime  int64  `json:"firsttime"`
-	Popularity int    `json:"popularity"`
+	Popularity int64  `json:"popularity"`
 	Devices    []struct {
-		Did        int    `json:"did"`
+		Did        int64  `json:"did"`
 		Objecttype string `json:"objecttype"`
 		Macaddress string `json:"macaddress,omitempty"`
 		Vendor     string `json:"vendor,omitempty"`
@@ -346,8 +346,8 @@ type EndpointDetails struct {
 		Lasttime  int64  `json:"lasttime"`
 	} `json:"ips"`
 	Locations []struct {
-		Latitude  int    `json:"latitude"`
-		Longitude int    `json:"longitude"`
+		Latitude  int64  `json:"latitude"`
+		Longitude int64  `json:"longitude"`
 		Country   string `json:"country"`
 		City      string `json:"city"`
 	} `json:"locations"`
